using tenant scoped roles and email domains for user role access control
